Add a BGZF reading stream
